Move `eventsourcing.Data` to `saga.EventedData`
This is in preparation for allowing use of evented data instances for CRUD sagas.
