+++ src/ax/saga/instance.go
@@ -3,7 +3,6 @@ package saga
 import (
 	"context"
 
-	"github.com/golang/protobuf/proto"
 	"github.com/jmalloc/ax/src/ax/ident"
 	"github.com/jmalloc/ax/src/ax/persistence"
 )
@@ -29,22 +28,6 @@ type Instance struct {
 
 // Revision is the version of a saga instance.
 type Revision uint64
-
-// Data is an interface for application-defined data associated with a saga
-// instance.
-type Data interface {
-	proto.Message
-
-	// SagaDescription returns a human-readable description of the saga
-	// instance.
-	//
-	// Assume that the description will be used inside log messages or displayed
-	// in audit logs.
-	//
-	// Follow the same conventions as for error messages:
-	// https://github.com/golang/go/wiki/CodeReviewComments#error-strings
-	SagaDescription() string
-}
 
 // InstanceRepository is an interface for loading and saving saga instances.
 type InstanceRepository interface {
